+++ kong/client_test.go
@@ -199,3 +199,30 @@ func TestTestWorkspace(T *testing.T) {
 	assert.NoError(err)
 	assert.Equal(currWorkspace.Config, origWorkspace.Config)
 }
+
+func TestBaseRootURL(t *testing.T) {
+	t.Run("default", func(t *testing.T) {
+		client, err := NewClient(nil, nil)
+		require.NoError(t, err)
+		require.NotNil(t, client)
+
+		require.Equal(t, client.BaseRootURL(), "http://localhost:8001")
+	})
+
+	t.Run("set via env", func(t *testing.T) {
+		t.Setenv("KONG_ADMIN_URL", "https://customkong.com")
+		client, err := NewClient(nil, nil)
+		require.NoError(t, err)
+		require.NotNil(t, client)
+
+		require.Equal(t, client.BaseRootURL(), "https://customkong.com")
+	})
+
+	t.Run("set via flag", func(t *testing.T) {
+		client, err := NewClient(String("https://customkong2.com"), nil)
+		require.NoError(t, err)
+		require.NotNil(t, client)
+
+		require.Equal(t, client.BaseRootURL(), "https://customkong2.com")
+	})
+}